requirements/requirements.txt
django==4.1.4
djangorestframework==3.14.0

django-cors-headers==3.13.0
django-countries==7.5
dj-database-url==0.5.0
django-phonenumber-field==7.0.0
django-storages==1.13.1

factory-boy==3.2.1
django-browser-reload==1.6.0 
django-compressor==4.1
django-tailwind==3.4.0
django-money==3.0.0

# Dates
pendulum==2.1.2

# CLI
click==8.1.3

# Postgres
psycopg2==2.9.3

# Tasks
celery[redis]==5.2.7
celery==5.2.7
django-celery-beat==2.4.0

# Environment variables
python-dotenv==0.15.0

# WSGI
gunicorn==20.0.4

# Static files
whitenoise==6.2.0

# Phone numbers
phonenumbers==8.13.0

# Short UUID
shortuuid==1.0.11

# Google
google-auth==2.16.0

# Billing
stripe==4.2.0